Split tests for starting processes into smaller files Sanity Checks: Performs tests for testing the behavior if supplied parameters are invalid or the process fails because of errors Security Checks: Tests any and all security related issues StartCallbackTypeTests: For each available callback type, the process instance handling is tested
